+++ packages/snap-preact-components/src/components/Atoms/Icon/Icon.stories.tsx
@@ -5,19 +5,24 @@ import { IconProps } from '../../../types';
 import { iconPaths } from './paths';
 
 import { componentArgs } from '../../../utilities';
+import { ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs/blocks';
+// @ts-ignore
+import Readme from '../Icon/readme.md';
 
 export default {
 	title: `Atoms/Icon`,
 	component: Icon,
 	parameters: {
+		docs: {
+			page: () => (
+				<div>
+					<Readme />
+					<ArgsTable story={PRIMARY_STORY} />
+				</div>
+			),
+		},
 		actions: {
 			disabled: true,
-		},
-		controls: {
-			disabled: false,
-		},
-		options: {
-			showPanel: true,
 		},
 	},
 	argTypes: {
@@ -144,6 +149,7 @@ Gallery.args = {
 };
 Gallery.parameters = {
 	controls: {
+		expanded: false,
 		disabled: true,
 	},
 	options: {