+++ packages/snap-preact-components/src/components/Atoms/Merchandising/InlineBanner.stories.tsx
@@ -4,10 +4,24 @@ import { InlineBanner, InlineBannerProps } from './InlineBanner';
 
 import { componentArgs } from '../../../utilities';
 import { Snapify } from '../../../utilities/snapify';
+import { ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs/blocks';
+
+// @ts-ignore
+import Readme from '../Merchandising/readme.md';
 
 export default {
 	title: `Atoms/InlineBanner`,
 	component: InlineBanner,
+	parameters: {
+		docs: {
+			page: () => (
+				<div>
+					<Readme />
+					<ArgsTable story={PRIMARY_STORY} />
+				</div>
+			),
+		},
+	},
 	argTypes: {
 		banner: {
 			description: 'InlineBanner content store reference',
@@ -18,6 +32,19 @@ export default {
 				},
 			},
 			control: { type: 'none' },
+		},
+		layout: {
+			description: 'Banner layout',
+			type: { required: true },
+			table: {
+				type: {
+					summary: 'string',
+				},
+			},
+			control: {
+				type: 'select',
+				options: ['grid', 'list'],
+			},
 		},
 		...componentArgs,
 	},
@@ -41,7 +68,7 @@ export default {
 const snapInstance = Snapify.search({ globals: { siteId: 'scmq7n', search: { query: { string: 'glasses' } } } });
 
 const Template = (args: InlineBannerProps, { loaded: { controller } }) => (
-	<InlineBanner banner={controller?.store?.merchandising?.content.inline[0]} />
+	<InlineBanner banner={controller?.store?.merchandising?.content.inline[0]} {...args} />
 );
 
 export const Default = Template.bind({});