+++ packages/snap-preact-components/src/components/Molecules/FacetHierarchyOptions/FacetHierarchyOptions.stories.tsx
@@ -5,10 +5,24 @@ import { FacetHierarchyOptions, FacetHierarchyOptionsProps } from './FacetHierar
 import { componentArgs } from '../../../utilities';
 import { Snapify } from '../../../utilities/snapify';
 import { FacetDisplay } from '../../../types';
+import { ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs/blocks';
+
+// @ts-ignore
+import Readme from '../FacetHierarchyOptions/readme.md';
 
 export default {
 	title: `Molecules/FacetHierarchyOptions`,
 	component: FacetHierarchyOptions,
+	parameters: {
+		docs: {
+			page: () => (
+				<div>
+					<Readme />
+					<ArgsTable story={PRIMARY_STORY} />
+				</div>
+			),
+		},
+	},
 	decorators: [
 		(Story) => (
 			<div style={{ maxWidth: '300px', border: '1px solid lightgrey', padding: '8px' }}>