+++ packages/snap-preact-components/src/components/Organisms/Autocomplete/Autocomplete.test.tsx
@@ -1,0 +1,8 @@
+import { h } from 'preact';
+import { render } from '@testing-library/preact';
+
+describe('Results Component', () => {
+	it('renders grid view', () => {
+		expect(true).toBe(true);
+	});
+});