+++ packages/snap-preact-components/src/types.ts
@@ -14,6 +14,7 @@ export interface Result {
 	id: string;
 	mappings: Mappings;
 	attributes: any;
+	type?: string;
 }
 export enum BannerType {
 	HEADER = 'header',
@@ -143,7 +144,7 @@ export interface BaseFacet {
 	display: FacetDisplay | any; //TODO: update mock data, 'any' to pass pagination tests
 	filtered: boolean;
 	label: string;
-	collapse: boolean;
+	collapsed: boolean;
 	toggleCollapse?: () => any; //TODO: update mock data, optional to pass pagination tests
 	controller?: any;
 }
@@ -204,7 +205,7 @@ export interface BaseFacetValue {
 	label: string;
 	count?: number;
 	url?: any;
-	value?:any;
+	value?: any;
 	custom?: any;
 }