+++ packages/snap-preact-demo/CHANGELOG.md
@@ -3,6 +3,103 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/searchspring/snap/compare/v0.1.0...v0.1.1) (2021-04-29)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+# 0.1.0 (2021-04-29)
+
+
+### Reverts
+
+* revert removing version from each package.json ([a0471de](https://github.com/searchspring/snap/commit/a0471dee9794c7044bd0231d645de3a831983a52))
+
+
+
+
+
+## 1.0.15 (2021-04-28)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 1.0.14 (2021-04-28)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 1.0.13 (2021-04-27)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 1.0.12 (2021-04-26)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 1.0.11 (2021-04-23)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 1.0.10 (2021-04-23)
+
+
+### Bug Fixes
+
+* icon gallery hiding controls ([#19](https://github.com/searchspring/snap/issues/19)) ([20a9d24](https://github.com/searchspring/snap/commit/20a9d246408a404fcfe2ca7b27541a8215e60f79))
+
+
+
+
+
+## 1.0.9 (2021-04-23)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
+## 1.0.8 (2021-04-23)
+
+
+### Bug Fixes
+
+* **husky:** migrate v4 to v6 ([#15](https://github.com/searchspring/snap/issues/15)) ([1d15c8f](https://github.com/searchspring/snap/commit/1d15c8f24467bc91b28039db51c35c02199c0774))
+
+
+
+
+
+## 1.0.7 (2021-04-22)
+
+**Note:** Version bump only for package @searchspring/snap-preact-demo
+
+
+
+
+
 ## 1.0.6 (2021-04-21)
 
 **Note:** Version bump only for package @searchspring/snap-preact-demo